refactor(McpHost): Transport関連クラスを専用フォルダに集約
- SseWriter, StreamableHttpTransport をプロジェクトルートから Transport/ に移動
- SseResponseHelper を Routing/ から Transport/ に移動
- 名前空間を RhinoMCPServer.McpHost.Transport に統一
